Unittest updates
Added factory method to create a SubtitleVersion.  When writing tests,
sometimes it's nice to avoid the workflow logic and just force a version
to be created.  Added a way to create the verion when calling VideoFactory.

Added utils.dates modules to handle a common issue with unittests --
controlling the value of now().  It also has some utility functions that
are useful for the billing reports.