Added subtitles_complete signal
This fires when a subtitle set is completed.  It's like
subtitles_published(), but it will fire regardless of if the subtitles
are public or not.

Basically it fires whenever subtitles_complete is set to True ...except
of course for tasks which has a bunch of special code.

Added method to mark subtitles complete and also emit the signal.

Added a system to suspend sending the event (and in theory other
events).  We need this because the mark_complete() is called by
in the middle of pipeline.add_subtitles().  We want to wait until the
end of all the work to emit the signal.

Refactored/simplified a few parts of pipeline.add_subtitles().