Docker build update
Created a non-root user to use when running our app/worker containers.
Docker security is pretty good, but they also recommend not using root
for extra permission.

Use /var/run/amara as the working directory when running our containers.
This is nice because it means any files created don't get copied into
the unisubs repo.  Renamed /opt/apps/unisubs to /opt/apps/amara.

Added .git to the .dockerignore.  This lowered the amount of data we
send to the docker daemon when building by about half (700M to 300M).
The only thing we were using it for was to create the commit file, so
I made a build script that updated that then called docker build.

Use 1 ADD command to copy the shell scripts into the docker image.
Before it was 1-per-script and that resulted in a bunch of extra image
layers.

Updated the requirements file to include some SSL-related improvements.
This fixes some warnings when running PIP and may also allow us to
handle google's certs correctly.

Don't set --no-cache when building the container.  The cache can speed
things up and I think it should be safe at this point.  Reordered the
Dockerfile so that slow operations that like apt-get install are towards
the top.  This makes it more likely that they get cached.