Changed way we keep the new video index updated
Before we would do it by having a needs_update field in the index table
and setting that to 1 when we wanted to update.  However, that was
causing deadlocks on staging.  I believe the reason for that is because
the index table is a MyISAM table and the other tables are InnoDB.
Mixing reads/writes to the two causes issues.  I'm not 100% sure how it
happened, but I believe it's something like this:

- A worker process wants to update the video index so it:
- A web process is doing a video search
- The worker process selects a bunch of rows from the videos/subtitles
  table to figure out what text should be in there. (Locking those
  rows)
- The web process does the search which is a join between the videos
  table and the video index table.  It locks the video index table for
  the select, but it can't finish it because of the above lock.
- The worker process now wants to update the video index table, so it
  runs an UPDATE statement.  However, it's blocked because of the lock
  from the web process.

Like I say, I'm not 100% sure what's happening, but i'm pretty sure it's
because of an interaction between the two tables.  So I changed it to go
back to the old system, we don't track what needs to be updated and
instead simply update the index in Video.update_search_index() which is
called inside the tasks.video_changed().