Log errors to elasticsearch
Catch errors in our middleware.  Format the stack trace in our logging
handler.

Added a module to format the function local variables for the stack
trace.  Switched to using that for the GET/POST variables as well.

Added view that can be used to test the error logging.  Remoed the
get_feedback view which we weren't using anymore.