Docker container fixes
Dropped the old system for finding the amara repo.  We use
/var/run/amara as our CWD in the docker container, but the code is in
/opt/apps/amara.  Before we would symlink a couple python scripts from
/opt/apps/amara to /var/run/amara, then monkey with the path in
startup.py using os.path.realpath().  However, that seemed to sometimes
fail to follow the symlink for unknown reasons.

Replaced it with a new system where we just setup the path in the
Dockerfile.  We also setup the MANAGE_SCRIPT env variable and use that
to find where manage.py is.  This seems to fix the issues of amara
randomly not starting up or failing to reload the code.

Don't change the entrypoint in our docker-compose file.  We can just use
regular commands to run the app/worker.