Added index page for staff pages.
We have a few of these pages now and I think it would be nice to have a
way to organize them and not force staff users to remember a multiple
urls if they want to do their work.

Remove ":" char from key-value tables.  I think it looks nicer without
it in many cases.  In the cases where we want a colon, we can manually
put it in.
foo