Added replacement for Video.get_or_create_for_url()
There were a bunch of issues for that one.  The main issue was that if
some attribute wasn't set inside the method, or the video wasn't added
to its team, then we would have a race condition.  Video.add() doesn't
have that issue.  It's also a lot cleaner than the old one and has some
tests