Reimplemented task-related activity
I did this without writing tests, or really implementing it cleanly.  I
just more-or-less replaced the old call with the new one.  This is
because I'm afraid to touch the tasks code and also because it's going
away soon, so I didn't feel like we should clean it up.

Made stubs for the member added/removed actions.  This makes the type
codes correct.