Fixed API tests
Added some monkey patching to bring in the compat module from the older version
of django-rest-framework.  This allows the tests to run again.

Fixed a bunch of minor issues with the tests.

Implemented the fix for the complete and send_back attributes on the
tasks API differently.  This one allows complete/send_back to work in
a PUT request.