Back to RTD for our API documentation
The main issue that moved us away was that we couldn't include
documentation from the amara-enterprise branch.  The new method should
let us do it, since we can update the API_DOCS setting in
amara-enterprise and include more modules to take the docstrings from.

Changed the document structure a bit.  Now all team-related API
resources are listed together.  To make this work, we moved all the
resource sections to directly below the API section.  All the other
stuff that used to be directly below is now under API->Overview.