Replaced the activity hack
Rather than messing with the select columns, just use FORCE INDEX to
tell mysql which index to use.  Also, move the code to
activity.models rather than having it burried in the template code.

Implemented a way to use FORCE INDEX with a new app.  We could have used
the django-mysql app for this, but that has a lot of code in it.  It
seemed like overkill for what we want.