Added extra margin to editor hover menu
Addresses Issue #1740

https://github.com/pculture/unisubs/issues/1740

An empty, invisible div was added to extend the hover area of the editor hover menu.
It acts as an extra margin of 45px to the left of the menu,
eliminating the gap between the icon and the hover menu,
where it is very easy to fall out of the hover area and lose the menu.

This margin runs the full lenght of the editor hover menu,
but can obviously be shortened to match the lenght of the tool icon
if the current area provides too much hover space.