Open adjacent cell after deleting a subtitle
Addresses Issue #2179

https://github.com/pculture/unisubs/issues/2179

Previously, after deleting a subtitle cell,
either by pressing "alt+backspace" or with the context menu,
the cursor would vanish along with the deleted cell.

With the cursor gone and no adjacent cell open for editing,
it was impossible for a user to repeat a press of "alt+backspace" to delete an additional cell.

With this commit, an adjacet cell (if it exists) will open for editing after a subtitle cell is deleted.
If "alt+backspace" is pressed in succession, adjacent cells will be deleted until no cells remain.

This behavior is not produced when deleting a subtitle cell using the context menu.

Note: I found it difficult to delete multiple cells in a row using "alt+backspace" even with this new functionality.
This is because unless the "alt" key is released rather quickly,
the operating system will pick up on the keypress and remove focus from the browser.