Fix for activity when deleting videos in a team
The issue was that when you delete a video, you also delete a team
video.  So our signal handler for removing the team video was firing
off and messing things up.  Changed it so that we have an explicit
signal for removing a team video and hook up to that.