Added a secure_id field to users
This is what we should use for our API responses (see #2625).

Display the secure ID on the admin listing.

Added a general way to handle secure IDs.  We might want to do this with
other classes as well.

Replaced old pycrypto with a newer version.

Added a memoize module that I inteded to use, but it didn't work out.
Still the code seem good to add.