Fixes/refactoring for sync_metadata
This fixes a couple issues:
  - Nothing was starting because there was an circular import issue
  - The tests weren't passing because we were passing
    primary_audio_languge_code to convert_language_code() even when it
    was blank.

I tried a couple fixes, but I decided the cleanest way would be to move
the sync_metadata() code out of google and into syncing.youtube.  I
think this works better with the general strategy that the syncing
module handles the high-level stuff and the google module handles the
low-level HTTP requests.

Also removed the error handling from sync_metadata().  We have
high-level error handling code that can catch errors.  That code also
does things like store a SyncingError in the database and allow us to
retry it.