Fixes #784. Added unit tests for PlaybackModeController.
Added mock for $timeout service with a helper for simulating
sequential resolution of promises in tandem with other time-based
services (e.g. the VideoPlayer).

Also removed logging from the PlaybackModeController in favor of
passing a reason string into the state machine reset() function
which can be inspected in a debugger, or logged manually later.