Allow deleting subtitle languages via the API (#2635)
Replaced teams.permissions.can_delete_language() with a permission on
the LanguageWorkflow class.  This is more flexible and can work with
non-team videos.

Updated the permissions code to require is_superuser rather than
is_staff.  I don't think we want to allow all staff users to do this.

Added API code and tests for it.