Updated team create API (pculture/amara-enterprise#744)
Allow partners to create teams with different workflow types via the
API.

Changed the TeamWorkflow class a bit to make this work.  Made it so the
type_code and api_code are set during the register() call, rather than
class attributes.  It's awkward to have api_code as a class attribute,
since it's optional and we subclass TeamWorkflows.  This causes problems
you subclass a TeamWorkflow with api_code set, and you don't explicitly
set it back to None in the subclass.

Updated documentation to describe creating teams.