Couple of notification fixes
Only call the on_video_added() handler if a team video is created, not
just updated.

Use a different method of checking when users are removed from teams.  I
think this one works better.

Use pre_delete rather than post_delete, this way if we want to do some
DB lookups or something (for example getting the project for a removed
video), then the data is available.