One more try to fix the messaging issue
The messages are getting created, but the notifications are not firing.
I really think it has something to do with how we calculate the ids to
send to send_new_messages_notifications().  Let's avoid bulk_create and
just create them one by one.

Updated the prepdeploy command so that it works better when you deploy
multiple times in 1 day.